fix(ci/cd): running os specific tests
